// frontend/lib/api/models/MonthlySessionCreate.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * Création d'une session mensuelle.
 */
export type MonthlySessionCreate = {
    /**
     * Premier jour du mois (ex: 2026-01-01)
     */
    month: string;
    /**
     * Ex: 'Janvier 2026'
     */
    name: string;
};